mdflow/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "mdflow"
version = "0.1.0"
description = "Configuration-driven multi-agent workflow system"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "CodeWeaver Team"}
]
keywords = ["workflow", "agent", "llm", "langgraph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "pydantic>=2.0.0",
    "python-frontmatter>=1.0.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
mdflow = "mdflow.cli:main"

[project.urls]
Homepage = "https://github.com/Honglin20/CodeWeaver"
Repository = "https://github.com/Honglin20/CodeWeaver"
Documentation = "https://github.com/Honglin20/CodeWeaver/blob/main/README.md"

[tool.setuptools]
packages = ["mdflow", "mdflow.core", "mdflow.workflows"]

[tool.setuptools.package-data]
mdflow = ["workflows/**/*.md"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
